requirements.txt
anyio==4.7.0
APScheduler==3.10.4
certifi==2024.12.14
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
